Pass along new tab / tab tap actions
